requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
croniter==2.0.1
idna==3.6
mysqlclient==2.2.4
numpy==1.26.4
pandas==1.4.1
pyodbc==5.1.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
python-dotenv==0.19.1
./packages/elindus_utils-1.0.5-py3-none-any.whl
